requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.10.0
pydantic-settings>=2.1.0

# OpenAI Agents SDK
openai-agents>=0.0.1

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.0
